chore(configs): rename flow-typed-install flowTypedInstall
